+++ src/features/dashboard/dashboard.tsx
@@ -1,47 +1,10 @@
 import { useStore } from 'effector-react';
 import React from 'react';
 
-import { DashboardSidebar } from './components/dashboard-sidebar';
-import {
-  $isListType,
-  $isLoading,
-  $noSearchResults,
-  SearchSortBlock,
-} from './components/search-sort-box';
-import { SearchResults } from './components/search-sort-results';
-
-const MainContent = () => (
-  /* countries-list--grid-view and countries-list--list-view for layout switching */
-  <div
-    className={`progress-dashboard__countries-list countries-list
-    ${
-      useStore($isListType)
-        ? 'countries-list--list-view'
-        : 'countries-list--grid-view'
-    }`}
-  >
-    {useStore($noSearchResults) ? null : (
-      <div className="countries-list__grid-header">
-        <div className="countries-list__grid-row">
-          <div className="countries-list__grid-col countries-list__grid-col--country">
-            Country name
-          </div>
-          <div className="countries-list__grid-col countries-list__grid-col--date">
-            Date of joining
-          </div>
-          <div className="countries-list__grid-col countries-list__grid-col--progress">
-            Progress
-          </div>
-          <div className="countries-list__grid-col countries-list__grid-col--schools">
-            Schools with connectivity
-          </div>
-          <div className="countries-list__grid-col countries-list__grid-col--link" />
-        </div>
-      </div>
-    )}
-    {useStore($isLoading) ? <p>Loading...</p> : <SearchResults />}
-  </div>
-);
+import { DashboardSidebar } from './dashboard-sidebar';
+import { $isListType, $isLoading, $noSearchResults } from './model';
+import { SearchSortBox } from './search-sort-box';
+import { SearchResults } from './search-sort-results';
 
 export const Dashboard = () => (
   <>
@@ -50,8 +13,36 @@ export const Dashboard = () => (
         <div className="progress-dashboard__row">
           <DashboardSidebar />
           <div className="progress-dashboard__countries-col">
-            <SearchSortBlock />
-            <MainContent />
+            <SearchSortBox />
+            <div
+              className={`progress-dashboard__countries-list countries-list
+                ${
+                  useStore($isListType)
+                    ? 'countries-list--list-view'
+                    : 'countries-list--grid-view'
+                }`}
+            >
+              {useStore($noSearchResults) ? null : (
+                <div className="countries-list__grid-header">
+                  <div className="countries-list__grid-row">
+                    <div className="countries-list__grid-col countries-list__grid-col--country">
+                      Country name
+                    </div>
+                    <div className="countries-list__grid-col countries-list__grid-col--date">
+                      Date of joining
+                    </div>
+                    <div className="countries-list__grid-col countries-list__grid-col--progress">
+                      Progress
+                    </div>
+                    <div className="countries-list__grid-col countries-list__grid-col--schools">
+                      Schools with connectivity
+                    </div>
+                    <div className="countries-list__grid-col countries-list__grid-col--link" />
+                  </div>
+                </div>
+              )}
+              {useStore($isLoading) ? <p>Loading...</p> : <SearchResults />}
+            </div>
           </div>
         </div>
       </div>
